Adds new package dependencies to Makefile (due to various rewrites of the code).
First of all, the ffpb-debug script is now written in lua and thus requires
a lua interpreter to be installed. But since the script makes use of some
gluon-specific lua-functions/-variables (such 'site_conf' in site_conf.lua),
we have to make ffpb-debug depend on the 'gluon-core', which in turn would
define all other dependencies.
Moreover, since ffpb-debug-live got rewritten to work with Barrier Breaker
(i.e. gluon 2014.4 and later), the already defined dependency towards the
busybox package does not suffice any longer, as we fetch interface-details
and -statistics via ubus now. As ubus outputs its data in JSON, we moreover
need the helper functions from the 'libubox' package for a convenient hand-
ling of such data in shell scripts.