Adds process-watchdog for A.L.F.R.E.D.
Monitors whether or not the A.L.F.R.E.D. is still running on a node or
not. Version 2013.x crashes from time to time and thus some nodes dis-
appear from the map. It is thus suitable to monitor the state of the
daemon on the node itself by periodically making requests to avoid un-
necessary support-requests and manual restarts of these nodes. In a
failure state, the daemon can simply be restarted by triggering its
init-script.

Please note that the bug is most likely gone in v2014.1 and newer and
and thus the package does not need to be included in the firmware in
newer firmwares.