Adds check for A.L.F.R.E.D major version > 2013 and sends ffpb-debug-report on crash.
As it is only assumed that the bug, which causes A.L.F.R.E.D. 2013.4 to
crash, got fixed in 2014.1, it is important to know whether or not this
really is the case. Hence, an additional check got implemented that
tests for the A.L.F.R.E.D. daemon's major version. If the major version
is greater than 2013 an additional ffpb-debug-report gets issued
automatically, so that the dev-team gets informed about the crash.