Adjusts ffpb-banner to avoid file collision error messages.
By now, the package was trying to overwrite "/etc/banner" in the image,
which caused opkg to stop the operation, as this file is already
provided by the base-files package. Without the --force-overwrite option
set for opkg (e.g. in the global gluon Makefile, where it is not set by
default), the installation of the package will fail.
Hence the problem got circumvented by renaming the customized banner file
provided by this package to "/etc/banner.ffpb" and using the prerm,
preinst, and postinst hooks of the packages' make-file to rename the
already existent "/etc/banner" to "/etc/banner.openwrt" and establish a
symbolic link "/etc/banner" to "/etc/banner.ffpb".
For deinstallation of the package, this link gets removed and a new one
(pointing at "/etc/banner.openwrt") gets created.