Adds the ne package ffpb-node-tuning to ffpb-packages.
As its name states, this packages applies some config optimizations
(like tweaking kernal knobs via sysctl) on the nodes.
For the most knobs Gluon sticks for OpenWRT defaults for good reasons,
as some values need to be set according to the environment of the
nodes (e.g. the network's size). As OpenWRT basically aims at provid-
ing a stable system for standalone WiFi access-points, the system's
default configuration may not be appropriate anymore the larger the
network grows, though.