Renames wr1043ndv2-2-stable package for a broader use.
The package by now was rather aimed to target on the migration of
TP-Link TL-WR1043NDv2 devices from a pseudo-stable firmware to the
officially stable version 0.6.0.

However, as it might become necessary to change other configuration
values from AA to be compliant with their corresponding values in BB
independent from the particular hardware-type, the package got renamed
to reflect this use. Moreover, as there only is the need to migrate
TP-Link TL-WR1043NDv2 devices once, with the new name the package can be
reused for the migration to Chaos Calmer by adapting the included
scripts.