Adds final exit status and treatment of return values from non-ffpb packages.
The file zz~-ffpb-upgrade will only be delete if its return value is 0.
For some reason we forgot to add an 'exit 0' at the end of the file if
everything is OK. This has been added.

To deal with invariant scripts that were not executed properly, we now
test the return value of do_component for these scripts as well. The
function do_component however may return early with a return code not
equal to 0, if a version for that package/script could not have been
determined. We now return the code 255 in this case and ignore that
state in "main", as we are not interested in non-ffpb packages.