Adds script to change uci's wireless config to be in BB format.
As mentioned in [1], the hwmodes "11ng" and "11na" are no longer valid
in BB and become "11g" and "11a". Enabling 802.11n modes are set via the
htmode option now.

Moreover, as the version of BB we rely on with Gluon v2014.4 comes with
a faulty setup script for mac80211 that prevents setting of htmode in
adhoc mode, we replace the default script with a newer one from [2].
Additionally, the mcast_rate value did not get applied correctly in the
BB base of v2014.4, which gets fixed by this exchange as well.

[1] https://dev.openwrt.org/ticket/17541
[2]
https://dev.openwrt.org/browser/trunk/package/kernel/mac80211/files/lib/netifd/wireless/mac80211.sh?rev=43853