Adds missing hostapd.sh replacement from CC to the package.
As it turned out, the old hostapd.sh from BB does not provide all the
necessary functions needed by the mac80211.sh that got added in commit
f2d788633930dfdc170802aad8b899f9d3ca7208. The corresponding hostapd.sh
from the CC trunk [1] however has this functions and works fine in BB
already. Hence, we add the netifd.sh from [1] into the package because
of the fixes in mac80211.sh in CC that did not get backported to BB
(yet). netifd.sh gets renamed to hostapd.sh by the Makefile for the
hostapd package and copied to /lib/netifd.

[1]
https://dev.openwrt.org/browser/trunk/package/network/services/hostapd/files/netifd.sh?rev=44438

Signed-off-by: Stefan Laudemann <thisco@webcake.de>