don't force 1043v2 routers to stable firmware any more
This has been necessary as we needed the 1043v2 models when no stable
firmware had had been released for them, yet. In order not to
"accidentially" force its users to a testing firmware we included this
now obsolete fix.

The file can be deleted as it is not part of the overlay or is listed
in sysupgrade.conf.