New Package ffho-site-auto-select
This package should automaticaly select the site of a gluon-node, if it is not done before.
Actualy 003-site-auto-select looks during firmware-upgrade, if a valid site is selected, else it sets a cron-job for ffho-site-auto-select.
ffho-site-auto-select use gluon-neighbour-info, to determine the site_code of the connected batman-neighbours (on interface != mesh-vpn). If it is a valid site, it is set as currentsite for this node, too.